requirements.txt
fastapi==0.75.0
uvicorn==0.17.6
aiofiles==0.8.0
python-multipart==0.0.5
google-cloud-pubsub==2.11.0